requirements.txt
fastapi
pydantic
pandas
sqlalchemy
uvicorn
pymysql
python-telegram-bot
google-genai
python-dotenv